--- GmodMongoDb/Binding/Annotating/NamespaceDoc.cs
@@ -1,12 +1,0 @@
-﻿using System.Runtime.CompilerServices;
-
-namespace GmodMongoDb.Binding.Annotating
-{
-    /// <summary>
-    /// Contains annotations to mark .NET members so they're exposed to Lua.
-    /// </summary>
-    [CompilerGenerated()]
-    class NamespaceDoc
-    {
-    }
-}
