--- GmodMongoDb/Binding/DataTransforming/NamespaceDoc.cs
@@ -1,12 +1,0 @@
-﻿using System.Runtime.CompilerServices;
-
-namespace GmodMongoDb.Binding.DataTransforming
-{
-    /// <summary>
-    /// Contains transformers that help convert between .NET and Lua types.
-    /// </summary>
-    [CompilerGenerated()]
-    class NamespaceDoc
-    {
-    }
-}
