--- GmodMongoDb/Binding/NamespaceDoc.cs
@@ -1,12 +1,0 @@
-﻿using System.Runtime.CompilerServices;
-
-namespace GmodMongoDb.Binding
-{
-    /// <summary>
-    /// Contains tools to bind .NET objects with Lua comfortably
-    /// </summary>
-    [CompilerGenerated()]
-    class NamespaceDoc
-    {
-    }
-}
