+++ GmodMongoDb/Binding/TypeTools.cs
@@ -1,11 +1,7 @@
-﻿using GmodMongoDb.Binding.Annotating;
-using GmodMongoDb.Binding.DataTransforming;
-using GmodNET.API;
+﻿using GmodNET.API;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Reflection;
-using System.Runtime.InteropServices;
 
 namespace GmodMongoDb.Binding
 {
@@ -15,204 +11,7 @@ namespace GmodMongoDb.Binding
     public static class TypeTools
     {
         private static readonly Dictionary<int, Type> MetaTableTypeIds = new();
-        private static readonly Dictionary<Type, Type> TransformerTypes = new();
-        private static string[] metaTableTypeNames = null;
 
-        /// <summary>
-        /// Iterates all Transformers and registers their Type for later use. In order to create a transformer you should inherit LuaValueTransformer
-        /// </summary>
-        internal static void DiscoverDataTransformers()
-        {
-            var transformerBaseType = typeof(BaseLuaValueTransformer);
-            var discoveredTypes = AppDomain.CurrentDomain.GetAssemblies()
-                .SelectMany(assembly => assembly.GetTypes())
-                .Where(type => type.IsClass && !type.IsAbstract && type.IsSubclassOf(transformerBaseType));
-
-            foreach (var transformerType in discoveredTypes)
-            {
-                var genericArguments = transformerType.BaseType.GetGenericArguments();
-
-                TransformerTypes.Add(genericArguments[0], transformerType);
-            }
-        }
-
-        /// <summary>
-        /// Eagerly search for metatable definitions and create the metatables.
-        /// </summary>
-        /// <param name="lua"></param>
-        internal static void CreateDiscoveredMetaTableDefinitions(ILua lua)
-        {
-            var metaTableBaseType = typeof(LuaMetaObjectBinding);
-            var discoveredTypes = AppDomain.CurrentDomain.GetAssemblies()
-                .SelectMany(assembly => assembly.GetTypes())
-                .Where(type => type.IsClass && !type.IsAbstract && type.IsSubclassOf(metaTableBaseType));
-            metaTableTypeNames = new string[discoveredTypes.Count()];
-            var i = 0;
-
-            foreach (var metaTableType in discoveredTypes)
-            {
-                var metaTableName = GetMetaTableTypeName(metaTableType);
-                var typeId = lua.CreateMetaTable(metaTableName);
-
-                if (!MetaTableTypeIds.ContainsKey(typeId))
-                    MetaTableTypeIds.Add(typeId, metaTableType);
-
-                // Have a place for this type's static methods
-                lua.CreateTable();
-                lua.Push(-1);
-                lua.Insert(1); // Move mt._static to the bottom of the stack so we can fill it later
-                // Add the static method table currently at the top of the stack to the global table
-                lua.PushSpecial(SPECIAL_TABLES.SPECIAL_GLOB);
-                lua.Insert(-2);
-                lua.SetField(-2, metaTableName);
-                metaTableTypeNames[i] = metaTableName;
-                lua.Pop(1); // pop the global table
-
-                MethodInfo[] methods = metaTableType.GetMethods();
-
-                foreach (var method in methods)
-                {
-                    var attributes = method.GetCustomAttributes<LuaMethodAttribute>();
-
-                    foreach (var attribute in attributes)
-                    {
-                        var stackPos = -2; // Position of the metatable
-                        attribute.PushFunction(lua, method, typeId);
-
-                        if (method.IsStatic)
-                        {
-                            stackPos = 1; // Position of the static table
-
-                            if (attribute.IsConstructor)
-                            {
-                                // TODO: Test what happens with overloaded initializers
-                                // Create a metatable for the static table and add the __call metamethod to it
-                                lua.CreateTable();
-                                attribute.PushFunction(lua, method);
-                                lua.SetField(-2, "__call");
-                                lua.SetMetaTable(stackPos); // Pops the metatable for the static table
-                            }
-                        }
-
-                        var methodName = attribute.Name ?? method.Name;
-                        lua.SetField(stackPos, methodName);
-                    }
-                }
-
-                PropertyInfo[] properties = metaTableType.GetProperties();
-
-                foreach (var property in properties)
-                {
-                    var attributes = property.GetCustomAttributes<LuaPropertyAttribute>();
-
-                    foreach (var attribute in attributes)
-                    {
-                        LuaPropertyAttribute.RegisterAvailableProperty(metaTableType, attribute.Name ?? property.Name, property.GetGetMethod()?.Name, property.GetSetMethod()?.Name);
-                    }
-                }
-
-                lua.Pop(1); // Pop the metatable
-                lua.Pop(1); // Pop the static functions table
-            }
-        }
-
-        /// <summary>
-        /// Cleans up the static tables made with <see cref="CreateDiscoveredMetaTableDefinitions"/>
-        /// </summary>
-        /// <param name="lua"></param>
-        internal static void CleanUpStaticFunctionTables(ILua lua)
-        {
-            if (metaTableTypeNames == null)
-                return;
-
-            lua.PushSpecial(SPECIAL_TABLES.SPECIAL_GLOB);
-
-            foreach (var metaTableName in metaTableTypeNames)
-            {
-                if (metaTableName == null)
-                    continue;
-
-                lua.PushNil();
-                lua.SetField(-2, metaTableName);
-            }
-
-            lua.Pop(1); // pop the global table
-        }
-
-        private static string GetMetaTableTypeName(Type type)
-        {
-            var classAttribute = type.GetCustomAttribute<LuaMetaTableAttribute>();
-            return classAttribute?.Name ?? type.Name;
-        }
-
-        /// <summary>
-        /// Generates userdata with a metatable by looking at attributes on the given object. Pushes the userdata onto the stack.
-        /// </summary>
-        /// <param name="lua"></param>
-        /// <param name="instance">The object to generate into userdata</param>
-        /// <returns>The metatable type id that was created for this userdata</returns>
-        public static int GenerateUserDataFromObject(ILua lua, LuaMetaObjectBinding instance)
-        {
-            int instanceTypeId = PushManagedObject(lua, instance);
-            
-            return instanceTypeId;
-        }
-
-        /// <summary>
-        /// Pushes a .NET managed object onto the stack as userdata.
-        /// 
-        /// Consider using <see cref="GenerateUserDataFromObject"/> if you want to interact with the userdata from Lua.
-        /// </summary>
-        /// <param name="lua"></param>
-        /// <param name="managed">The object to convert to userdata and push</param>
-        /// <returns>Returns the metatable type id which was applied to the userdata</returns>
-        public static int PushManagedObject(ILua lua, object managed)
-        {
-            var handle = GCHandle.Alloc(managed, GCHandleType.Weak);
-            ReferenceManager.Add(handle);
-
-            // CreateMetaTable will automatically find an existing metatable with this type name, so no need to look in MetaTableTypeIds for it. 
-            string typeName = GetMetaTableTypeName(managed.GetType());
-            var typeId = lua.CreateMetaTable(typeName);
-            lua.Pop(1);
-
-            if(managed is LuaMetaObjectBinding objectBinding)
-                objectBinding.MetaTableTypeId = typeId;
-
-            lua.PushUserType((IntPtr)handle, typeId);
-
-            return typeId;
-        }
-
-#nullable enable
-        /// <summary>
-        /// Pull userdata from the stack at the given position that has the given type id. Converts it to a .NET object.
-        /// </summary>
-        /// <param name="lua"></param>
-        /// <param name="typeId">The metatable type id of the userdata as returned by <see cref="PushManagedObject"/> or <see cref="GenerateUserDataFromObject"/></param>
-        /// <param name="stackPos">The position of the managed object</param>
-        /// <param name="forceKeepOnStack">Keep the object on the stack, false to remove it</param>
-        /// <returns>The .NET object converted from Lua</returns>
-        /// <exception cref="NullReferenceException">Thrown when a null pointer is found or the userdata is no longer valid</exception>
-        public static object PullManagedObject(ILua lua, int typeId, int stackPos = -1, bool forceKeepOnStack = false)
-        {
-            var handle = lua.GetUserType(stackPos, typeId);
-
-            if(!forceKeepOnStack)
-                lua.Remove(stackPos); // Remove the object
-
-            if (handle == IntPtr.Zero)
-                throw new NullReferenceException("Invalid instance pointer!");
-
-            object? reference = GCHandle.FromIntPtr(handle).Target;
-
-            if (reference == null)
-                throw new NullReferenceException("Userdata reference has gone away!");
-
-            return reference;
-        }
-#nullable disable
-        
         /// <summary>
         /// Push a value of a specific type to the Lua stack. Applies a transformer where possible.
         /// </summary>
@@ -253,20 +52,9 @@ namespace GmodMongoDb.Binding
 
                 return 1;
             }
-            else if (type == typeof(LuaFunctionReference) || type == typeof(LuaReference))
-            {
-                var reference = value as LuaReference;
-                reference.Push();
-
-                return 1;
-            }
-            else if (TransformerTypes.ContainsKey(type))
-                return lua.ApplyTransformerConvert(TransformerTypes[type], value);
             else
                 // TODO
                 throw new NotImplementedException($"This type is not registered for conversion to Lua from .NET! Consider building a Transformer. Type is: {type.FullName}");
-
-            return 0;
         }
 
         /// <summary>
@@ -290,35 +78,6 @@ namespace GmodMongoDb.Binding
                 || type == typeof(float)
                 || type == typeof(double))
                 value = lua.GetNumber(stackPos);
-            else if (type == typeof(LuaFunctionReference))
-            {
-                value = new LuaFunctionReference(lua, stackPos, forceKeepOnStack);
-                pop = false;
-            }
-            else if (type == typeof(LuaTableReference))
-            {
-                value = new LuaTableReference(lua, stackPos, forceKeepOnStack);
-                pop = false;
-            }
-            else if (type == typeof(LuaReference))
-            {
-                value = new LuaReference(lua, stackPos, forceKeepOnStack);
-                pop = false;
-            }
-            // Handle managed objects before trying the transformers
-            else if (TryGetMetaTableType(lua.GetType(stackPos), out Type metaTableType) && metaTableType == type)
-            {
-                value = PullManagedObject(lua, lua.GetType(stackPos), stackPos, forceKeepOnStack);
-                pop = false;
-            }
-            // Find a transformer to do the work
-            else if (TransformerTypes.ContainsKey(type))
-            {
-                var transformerType = TransformerTypes[type];
-                if (!lua.ApplyTransformerParse(transformerType, out value, stackPos, forceKeepOnStack))
-                    throw new InvalidCastException($"Could not apply transformer {transformerType} to get value from Lua stack at stackpos {stackPos}");
-                pop = false;
-            }
             else
                 // TODO
                 throw new NotImplementedException($"This type is not registered for conversion from Lua to .NET! Consider building a Transformer. Type is: {type.FullName}");
@@ -390,10 +149,8 @@ namespace GmodMongoDb.Binding
                 TYPES.NUMBER => typeof(double),
                 TYPES.STRING => typeof(string),
                 TYPES.BOOL => typeof(bool),
-                TYPES.FUNCTION => typeof(LuaFunctionReference),
-                TYPES.TABLE => typeof(LuaTableReference),
                 TYPES.NIL => null,
-                _ => typeof(LuaReference),
+                _ => throw new NotImplementedException($"This type is not registered for conversion from Lua to .NET! Type is: {luaType}"),
             };
         }
