--- GmodMongoDb/NamespaceDoc.cs
@@ -1,12 +1,0 @@
-﻿using System.Runtime.CompilerServices;
-
-namespace GmodMongoDb
-{
-    /// <summary>
-    /// Contains the userdata definitions that expose the .NET MongoDB Driver to Lua.
-    /// </summary>
-    [CompilerGenerated()]
-    class NamespaceDoc
-    {
-    }
-}
